test(autopipeline): consolidate full-duplex tests into one file
Merge the per-topic full-duplex test files (limiter, retry, otel, himport,
close, hookdeadlock) into autopipeline_fullduplex_test.go and remove them.
No test logic changes; the imports are unioned (adds internal/otel).
